made split between config and config loader
