+++ HABApp/config/__init__.py
@@ -1,2 +1,3 @@
-from .config import setup_config, CONFIG, InvalidConfigException, Config
 from .config import Openhab, Mqtt
+from .config import CONFIG
+from .config_loader import ConfigLoader, InvalidConfigException
