+++ HABApp/core/events/habapp_events.py
@@ -1,5 +1,8 @@
 from pathlib import Path
 
+TOPIC_RULES = 'HABApp.Rules'
+TOPIC_PARAM = 'HABApp.Parameters'
+
 
 class RequestFileLoadEvent:
     """Request (re-) loading of the specified file
