+++ HABApp/util/__init__.py
@@ -1,4 +1,4 @@
-from .wrapper import log_exception
+from .wrapper import log_exception, ignore_exception
 
 from .timeframe import TimeFrame
